+++ energyusage/utils.py
@@ -205,11 +205,12 @@ def log(*args):
 
     elif args[0] == "Final Readings":
         newline()
-        baseline_average, process_average, timedelta = args[1], args[2], args[3]
+        baseline_average, process_average, difference_average, timedelta = args[1], args[2], args[3], args[4]
         delete_last_lines()
         log_header(args[0])
         sys.stdout.write("{:<25} {:>48.2f} {:5<}\n".format("Average baseline wattage:", baseline_average, "watts"))
-        sys.stdout.write("{:<25} {:>48.2f} {:5<}\n".format("Average process wattage:", process_average, "watts"))
+        sys.stdout.write("{:<25} {:>48.2f} {:5<}\n".format("Average total wattage:", process_average, "watts"))
+        sys.stdout.write("{:<25} {:>48.2f} {:5<}\n".format("Average process wattage:", difference_average, "watts"))
         sys.stdout.write("{:<17} {:>62}\n".format("Process duration:", timedelta))
 
     elif args[0] == "Energy Data":
@@ -245,7 +246,7 @@ def log(*args):
     elif args[0] == "Assumed Carbon Equivalencies":
         log_header('Assumed Carbon Equivalencies')
         sys.stdout.write("{:<14} {:>65}\n".format("Coal:", ".3248635 kg CO2/kWh"))
-        sys.stdout.write("{:<14} {:>65}\n".format("Oil/Petroleum:", ".23 kg CO2/kWh"))
+        sys.stdout.write("{:<14} {:>65}\n".format("Petroleum:", ".23 kg CO2/kWh"))
         sys.stdout.write("{:<14} {:>65}\n".format("Natural gas:", ".0885960 kg CO2/kwh"))
 
     else: