+++ changelog.md
@@ -2,4 +2,5 @@
 
 v0.0.2
 * removed emissions from return of `evaluate`, now just returns value of user function
+* added `Average total wattage` to report to distinguish between baseline+process and just process
 * added to README to make it comprehensive